+++ QBCore_2F0666.base/resources/[qb]/qb-policejob/server/main.lua
@@ -154,7 +154,9 @@ RegisterNetEvent('police:server:policeAlert', function(text)
     for _, v in pairs(players) do
         if v and v.PlayerData.job.type == 'leo' and v.PlayerData.job.onduty then
             local alertData = { title = Lang:t('info.new_call'), coords = { x = coords.x, y = coords.y, z = coords.z }, description = text }
-            TriggerClientEvent('qb-phone:client:addPoliceAlert', v.PlayerData.source, alertData)
+            if GetResourceState('qb-phone') == 'started' then
+                TriggerClientEvent('qb-phone:client:addPoliceAlert', v.PlayerData.source, alertData)
+            end
             TriggerClientEvent('police:client:policeAlert', v.PlayerData.source, coords, text)
         end
     end
@@ -234,7 +236,9 @@ RegisterNetEvent('police:server:SendTrackerLocation', function(coords, requestId
         description = msg
     }
     TriggerClientEvent('police:client:TrackerMessage', requestId, msg, coords)
-    TriggerClientEvent('qb-phone:client:addPoliceAlert', requestId, alertData)
+    if GetResourceState('qb-phone') == 'started' then
+        TriggerClientEvent('qb-phone:client:addPoliceAlert', requestId, alertData)
+    end
 end)
 
 -- Threads